+++ eslint.config.js
@@ -2,11 +2,4 @@
 
 const init = require('eslint-config-metarhia');
 
-module.exports = [
-  ...init,
-  {
-    rules: {
-      'no-empty': ['error', { allowEmptyCatch: true }],
-    },
-  },
-];
+module.exports = [...init];